折り線の点選択のカスタムフック
