+++ src/components/OrigamiPost/hooks/useSelectPoints/index.tsx
@@ -1,0 +1,102 @@
+/* 
+step1での、選択した点を管理するカスタムフック
+**/
+import React, { useState, useEffect } from "react";
+import * as THREE from "three";
+import { renderPoint } from "../../logics/renderPoint";
+import { Point } from "@/types/three";
+
+type UseSelectPoints = (props: {
+  canvasRef: React.MutableRefObject<HTMLCanvasElement | null>;
+  sceneRef: React.MutableRefObject<THREE.Scene | null>;
+  cameraRef: React.MutableRefObject<THREE.PerspectiveCamera | null>;
+  rendererRef: React.MutableRefObject<THREE.WebGLRenderer | null>;
+  raycasterRef: React.MutableRefObject<THREE.Raycaster | null>;
+  inputStep: string;
+}) => {
+  selectedPoints: Point[];
+};
+
+export const useSelectPoints: UseSelectPoints = ({
+  canvasRef,
+  sceneRef,
+  cameraRef,
+  rendererRef,
+  raycasterRef,
+  inputStep,
+}) => {
+  const [selectedPoints, setSelectedPoints] = useState<Point[]>([]);
+
+  // モデル上の点を選択できるようにする処理
+  useEffect(() => {
+    const sizes = {
+      width: window.innerWidth - 320,
+      height: window.innerHeight,
+    };
+    if (inputStep !== "axis") return;
+    const canvas = canvasRef.current!;
+    const scene = sceneRef.current!;
+    const renderer = rendererRef.current!;
+    const camera = cameraRef.current!;
+    const raycaster = raycasterRef.current!;
+
+    const clickListener = (event: MouseEvent) => {
+      const mouse = new THREE.Vector2();
+      mouse.x = (event.clientX / sizes.width) * 2 - 1;
+      mouse.y = -(event.clientY / sizes.height) * 2 + 1;
+
+      // Raycasterのセットアップ
+      raycaster.setFromCamera(mouse, camera);
+
+      // エッジに対してRaycasterを適用して交差を調べる
+      const edges = scene.children.filter(
+        (child) => child.type === "LineSegments"
+      );
+      const intersects = raycaster.intersectObjects(edges, true);
+
+      if (intersects.length > 0) {
+        const point = intersects[0].point; // 最初の交差点の座標を取得
+
+        // 現在記録している点の数が2個未満の場合は、新しい点を追加, 2個以上の場合は最初の点を消して、新しい点を追加
+        let newPoints = [...selectedPoints];
+
+        if (selectedPoints.length < 2) {
+          newPoints.push([point.x, point.y, point.z]);
+        } else {
+          newPoints = newPoints.slice(1);
+          newPoints.push([point.x, point.y, point.z]);
+        }
+
+        // 既存のpointを削除する
+        scene.children = scene.children.filter(
+          (child) => child.name !== "Point"
+        );
+        // pointsを描画し直す
+        newPoints.forEach((point) => {
+          renderPoint({ scene, point });
+        });
+        setSelectedPoints(newPoints);
+
+        renderer.render(scene, camera);
+      }
+    };
+
+    canvas.addEventListener("click", clickListener);
+
+    return () => {
+      canvas.removeEventListener("click", clickListener);
+    };
+  }, [
+    selectedPoints,
+    inputStep,
+    canvasRef,
+    sceneRef,
+    rendererRef,
+    cameraRef,
+    raycasterRef,
+  ]);
+
+  return {
+    selectedPoints,
+  };
+};
